The code repair task is configured not to be manually deleted by the user

+++ services/abilitymgr/src/mission_info_mgr.cpp
@@ -194,7 +194,7 @@ bool MissionInfoMgr::DeleteAllMissionInfos(const std::shared_ptr<MissionListener
 
     for (auto listIter = missionInfoList_.begin(); listIter != missionInfoList_.end();) {
         if (!((listIter->missionInfo.lockedState) || (abilityMs_->IsBackgroundTaskUid(listIter->uid) ||
-            (listIter->missionInfo.unclearable)))) {
+            ((listIter->missionInfo.unclearable) && (listIter->missionInfo.runningState == 0))))) {
             missionIdMap_.erase(listIter->missionInfo.id);
             taskDataPersistenceMgr_->DeleteMissionInfo(listIter->missionInfo.id);
             if (listenerController) {
